feat(Exchange): move expired option market handling to the base exchange class
